Set up NavMesh
Created two planes that will act as test cells, added a parent object that holds a NavMesh surface for them both.

Made a couple of materials with slightly different colors so that we can differentiate between cells.

Created a capsule to act as a test unit. Added the NavMesh agent property to it.

Made an empty UnitManager object with a script that lets you set the destination of the test unit with a click. This is just for testing purposes.

Next step is to make a script for the cells so that they have a list of all units that are currently occupying them.