Player Square Detection
Essentially, this goes through all destinations tagged with "square". Then it checks to see if playerControlled is enabled. Then it checks for the closest distance squared (because performance). And finally it sets that unit's destination as the one closest to it. To see this example in action click player then space bar. This allows dynamic changes to the map, so it doesn't necessarily have to be limited to a 3x3, it doesn't even have to be a square.